// src/world/engine/hybrid/ruffle.d.ts
interface BaseLoadOptions {
    allowScriptAccess?: boolean;
    parameters?: URLSearchParams | string | Record<string, string> | null;
    autoplay?: AutoPlay;
    backgroundColor?: string | null;
    letterbox?: Letterbox;
    unmuteOverlay?: UnmuteOverlay;
    upgradeToHttps?: boolean;
    compatibilityRules?: boolean;
    favorFlash?: boolean;
    warnOnUnsupportedContent?: boolean;
    logLevel?: LogLevel;
    showSwfDownload?: boolean;
    contextMenu?: ContextMenu | boolean;
    preloader?: boolean;
    splashScreen?: boolean;
    maxExecutionDuration?: Duration;
    base?: string | null;
    menu?: boolean;
    salign?: string;
    forceAlign?: boolean;
    quality?: string;
    scale?: string;
    forceScale?: boolean;
    frameRate?: number | null;
    wmode?: WindowMode;
    playerVersion?: number | null;
    preferredRenderer?: RenderBackend | null;
    publicPath?: string | null;
    polyfills?: boolean;
    openUrlMode?: OpenURLMode;
    allowNetworking?: NetworkingAccessMode;
}

interface URLLoadOptions extends BaseLoadOptions {
    url: string;
}

interface DataLoadOptions extends BaseLoadOptions {
    data: Iterable<number>;
    swfFileName?: string;
}

interface MovieMetadata {
    readonly width: number;
    readonly height: number;
    readonly frameRate: number;
    readonly numFrames: number;
    readonly swfVersion: number;
    readonly backgroundColor: string | null;
    readonly isActionScript3: boolean;
    readonly uncompressedLength: number;
}
// ...
export interface RufflePlayer extends HTMLElement {
    onFSCommand: ((command: string, args: string) => boolean) | null;
    config: URLLoadOptions | DataLoadOptions | object;
    get readyState(): number;
    get metadata(): MovieMetadata | null;
    getExtensionConfig(): Record<string, unknown>;
    load(
        options: string | URLLoadOptions | DataLoadOptions
    ): Promise<void>;
    play(): void;
    get isPlaying(): boolean;
    volume: number;
    get fullscreenEnabled(): boolean;
    get isFullscreen(): boolean;
    setFullscreen(isFull: boolean): void;
    enterFullscreen(): void;
    exitFullscreen(): void;
    downloadSwf(): Promise<void>;
    pause(): void;
    set traceObserver(observer: ((message: string) => void) | null);
    PercentLoaded(): number;
}